Fix push hook (#20)
* Re-push produces meaningless version when push fails
Fixes #19

Need  update git-scripts to new version
link git-script/pr: https://github.com/zixia/git-scripts/pull/13

* style:

* ci: fix smoke-test

* 0.6.3

* refactor: Comment out unused code

* 0.6.4

* test: fix test error

* 0.6.5

* ci: :green_heart: update nodejs version

* 0.6.6

* fix: fix error on old version for git-scripts

* 0.6.7

* fix: :bug: fix npm bump again

* 0.6.8

* build: :arrow_up: update git-scripts to 0.5.3

* 0.6.9

* feat: :sparkles: suppport multi branch in pre-push

* 0.6.10

* fix:

* 0.6.11

* fix: :green_heart: fix error in test:pack

* 0.6.12

* chore: :arrow_up: upgrade git-scripts to 0.6.0

* 0.7.0

* fix: :bug: fix read properties error

* 0.7.1

+++ bin/pre-push.ts
@@ -14,6 +14,32 @@ import shell from 'shelljs'
 const NO_HOOK_VAR = 'NO_HOOK'
 const INNER_PRE_HOOK = 'CHATIE_INNER_PRE_HOOK'
 
+const argv = process.argv.slice(2)
+const remoteName = argv[0] || ''
+// const remoteUrl = argv[1] || ''
+
+interface RefInfo {
+  localBranch: string
+  localCommit: string
+  remoteBranch: string
+  remoteCommit: string
+}
+
+const refs: Array<RefInfo> = []
+
+for (let i = 2; i + 4 <= argv.length;) {
+  const ref: RefInfo = {
+    localBranch : argv[i++] || '',
+    localCommit : argv[i++] || '',
+    remoteBranch : argv[i++] || '',
+    remoteCommit : argv[i++] || '',
+  }
+  if (ref.localCommit.match(/^0+$/)) {
+    ref.localBranch = ''
+  }
+  refs.push(ref)
+}
+
 if (process.env[NO_HOOK_VAR]) {
   // user set NO_HOOK=1 to prevent this hook works
   process.exit(0)
@@ -24,11 +50,26 @@ if (process.env[INNER_PRE_HOOK]) {
   process.exit(0)
 }
 
+if (refs?.[0]?.localCommit.match(/^0+$/)) {
+  // delete remote branch
+  process.exit(0)
+}
+
+const packageVersion = require('../package.json').version
+const lastCommitMsg = shell.exec('git log --pretty=format:"%s" HEAD^0 -1', { silent : true }).stdout
+
+if (packageVersion === lastCommitMsg) {
+  process.exit(0)
+}
+
 shell.exec('npm run lint').code === 0 || process.exit(1)
 shell.rm('-f', 'package-lock.json')
 shell.exec('npm version patch --no-package-lock').code === 0 || process.exit(1)
 process.env[INNER_PRE_HOOK] = '1'
-shell.exec('git push').code === 0 || process.exit(1)
+
+const refMaps = refs.map(ref => ref.remoteBranch ? ref.localBranch + ':' + ref.remoteBranch : '')
+const cmd = ['git push', remoteName, ...refMaps].join(' ')
+shell.exec(cmd).code === 0 || process.exit(1)
 
 console.info(String.raw`
 ____ _ _        ____            _
